LangChain_Chatbot_OpenAI_Ollama/requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-core>=0.2.22
langchain-openai>=0.1.14
langchain-community>=0.2.9
openai>=1.40.0

